refactored headers into header component to reduce code duplication
